rename components, rebase project structure, fix styles
