+++ AlexanderMandrov/React/Messenger/src/components/App/App.scss
@@ -1,6 +1,6 @@
 @import '../../css/variables';
 
-.Layout {
+.App {
     &-main {
         margin-left: $drawerWidth;
     }
