--- AlexanderMandrov/React/Messenger/src/components/CreateMessage/index.js
@@ -1,3 +1,0 @@
-import CreateMessage from './CreateMessage';
-
-export default CreateMessage;
