+++ AlexanderMandrov/React/Messenger/src/components/MessageForm/MessageForm.scss
@@ -1,8 +1,9 @@
 @import '../../css/variables';
 
-.CreateMessage {
+.MessageForm {
     background-color: #fff;
     position: absolute;
     bottom: 0;
     width: calc(100% - #{$drawerWidth + 30});
+    max-height: calc(100vh - #{$offsetHeight}/2);
 }
