+++ AlexanderMandrov/React/Messenger/src/components/MessageForm/index.js
@@ -1,0 +1,3 @@
+import MessageForm from './MessageForm';
+
+export default MessageForm;
